# internal/handlers/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ShieldNet 360 Access Platform API",
    "description": "HTTP surface for the ZTNA access platform (see docs/architecture.md \u00a72). Covers access requests, reviews, grants, policy drafting / simulation, AI explain / suggest, and inbound SCIM (auto-sync users).\n\nThis spec is regenerated by `scripts/generate-swagger.sh` whenever a handler-level change ships (see CONTRIBUTING.md).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin server. The access API is served by cmd/ztna-api."
    }
  ],
  "paths": {
...
    "/workspace/policy/{id}": {
      "get": {
        "summary": "Get policy",
        "tags": [
          "policy"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/workspace/policy/{id}/simulate": {
      "post": {
        "summary": "Simulate a policy",
        "tags": [
          "policy"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/workspace/policy/{id}/promote": {
      "post": {
        "summary": "Promote a draft policy to active",
        "tags": [
          "policy"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/workspace/policy/{id}/diff": {
      "get": {
        "summary": "Diff a draft policy against the workspace's current live state",
        "description": "Returns a structured before/after report comparing the live baseline (union of OTHER live policies in the workspace) against the projection of the draft. Requires a previously-simulated draft; otherwise returns 409 Conflict.",
        "tags": [
          "policy"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "workspace_id required"
          },
          "409": {
            "description": "Draft has not been simulated or is no longer a draft"
          }
        }
      }
    },
...
  },
...
}